+++ stylelint.config.js
@@ -52,7 +52,15 @@ export default {
 				`flow-root`,
 			],
 		},
-		"declaration-property-value-no-unknown": true,
+		"declaration-property-value-no-unknown": [
+			true,
+			// See: https://github.com/stylelint/stylelint/issues/7241
+			{
+				typesSyntax: {
+					color: `| oklch( [ <percentage> | <number> | none] [ <percentage> | <number> | none] [ <hue> | none] [ / [<alpha-value> | none] ]? )`,
+				},
+			},
+		],
 		"font-family-name-quotes": `always-where-required`,
 		"font-family-no-duplicate-names": [
 			true,