+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and 
 
 - The minimum required version of `stylelint` has been increased to `16.8.1`.
 
+### Fixed
+
+- False positives of `declaration-property-value-no-unknown` rule for `oklch()`.
+
 ## [3.0.0] — 2024–06–07
 
 ### Changed